ulux-binding: Merge abstract test classes
