+++ bundles/binding/org.openhab.binding.ulux.test/src/test/java/org/openhab/binding/ulux/internal/handler/AudioVolumeMessageHandlerTest.java
@@ -3,11 +3,11 @@ package org.openhab.binding.ulux.internal.handler;
 import static org.mockito.Mockito.verify;
 
 import org.junit.Test;
-import org.openhab.binding.ulux.internal.ump.messages.AudioVolumeMessage;
+import org.openhab.binding.ulux.internal.ump.AbstractUluxMessageTest;
 import org.openhab.core.library.items.DimmerItem;
 import org.openhab.core.library.types.DecimalType;
 
-public class AudioVolumeMessageHandlerTest extends AbstractHandlerTest<AudioVolumeMessage> {
+public class AudioVolumeMessageHandlerTest extends AbstractUluxMessageTest {
 
 	@Test
 	public void testNoConfig() throws Exception {
