+++ bundles/binding/org.openhab.binding.ulux.test/src/test/java/org/openhab/binding/ulux/internal/handler/PageIndexMessageHandlerTest.java
@@ -3,11 +3,11 @@ package org.openhab.binding.ulux.internal.handler;
 import static org.mockito.Mockito.verify;
 
 import org.junit.Test;
-import org.openhab.binding.ulux.internal.ump.messages.PageIndexMessage;
+import org.openhab.binding.ulux.internal.ump.AbstractUluxMessageTest;
 import org.openhab.core.library.items.SwitchItem;
 import org.openhab.core.library.types.DecimalType;
 
-public class PageIndexMessageHandlerTest extends AbstractHandlerTest<PageIndexMessage> {
+public class PageIndexMessageHandlerTest extends AbstractUluxMessageTest {
 
 	@Test
 	public void testNoConfig() throws Exception {
