--- bundles/binding/org.openhab.binding.ulux.test/src/test/java/org/openhab/binding/ulux/internal/ump/AbstractMessageTest.java
@@ -1,16 +1,0 @@
-package org.openhab.binding.ulux.internal.ump;
-
-import java.nio.ByteBuffer;
-
-import org.openhab.binding.ulux.internal.AbstractUluxTest;
-
-public abstract class AbstractMessageTest extends AbstractUluxTest {
-
-	protected UluxMessageParser parser = new UluxMessageParser();
-
-	@SuppressWarnings("unchecked")
-	protected <T extends UluxMessage> T parseMessage(ByteBuffer buffer) {
-		return (T) parser.parseNextMessage(buffer);
-	}
-
-}
